pyproject.toml
[project]
name = "coding-agent"
version = "1.0.0"
description = "Production-ready AI coding agent with knowledge graph"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "User", email = "user@example.com"}
]

dependencies = [
    # Core LLM & Agent
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-ollama>=0.2.0",
    "langchain-groq>=0.2.0",
    "langchain-anthropic>=0.2.0",
    # Tracing & Observability
    "langfuse>=2.0.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "structlog>=24.1.0",
    "prometheus-client>=0.19.0",
    # Database - PostgreSQL
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.0",
    # Vector DB - Qdrant (RAG)
    "qdrant-client>=1.7.0",
    "langchain-qdrant>=0.1.0",
    "sentence-transformers>=2.2.2",
    "pypdf>=3.0.0",
    # API Gateway
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "websockets>=12.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "pydantic>=2.5.0",
    "aiofiles>=23.2.1",
    "orjson>=3.9.10",
    "cryptography>=42.0.0",
    # Sandboxing
    "docker>=7.0.0",
    # UI
    "streamlit>=1.40.0",
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "faker>=22.0.0",
    "locust>=2.20.0"
]

[project.scripts]
coding-agent = "src.interfaces.cli.main:app"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term"
